Cwe 190coverity warnings of this type adressed. Some where dismissed and maybe with reason but it seemed possible to remove them and hence obligatory ;p

* pr/1057:
  move back to original contract of isNetworksOverlap()
  Changed the behavior of methods that use NetUtils.cidrToLong(String)
  CWE-190 unit test for extremes of long netMaskFromCidr(long)
  CWE-190 netmask as long form cidr-size as method
  CID-1116482 cidrToLong cleanup of bitshift problem
  CID-1116483 cidr to netmask bitshifts guarded with casts
  CID-1116484 cast to long and use long as cidrsize type  and simpel test
  CID-1116485: cast cidr during bit shifting  and simple test included
  CID-1175714 casts before bit shift

+++ services/console-proxy-rdp/rdpconsole/src/main/java/streamer/ByteBuffer.java
@@ -418,7 +418,10 @@ public class ByteBuffer {
     }
 
     protected static long calculateUnsignedInt(byte value1, byte value2, byte value3, byte value4) {
-        return (calculateUnsignedByte(value1) << 24) + (calculateUnsignedByte(value2) << 16) + (calculateUnsignedByte(value3) << 8) + calculateUnsignedByte(value4);
+        return (((long)calculateUnsignedByte(value1)) << 24)
+                + (((long)calculateUnsignedByte(value2)) << 16)
+                + (((long)calculateUnsignedByte(value3)) << 8)
+                + calculateUnsignedByte(value4);
     }
 
     /**
